+++ src/Models/BranchAccessType.php
@@ -3,6 +3,7 @@
 namespace ItDelmax\AuthCache\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use ItDelmax\AuthCache\Models\BranchPartnerAccess;
 
 class BranchAccessType extends Model
 {
@@ -15,5 +16,11 @@ class BranchAccessType extends Model
     parent::__construct($attributes);
 
     $this->connection = config('auth_cache.connection') ?: parent::getConnectionName();
+  }
+
+
+  public function branchPartnerAccess()
+  {
+    return $this->hasMany(BranchPartnerAccess::class, 'BRANCH_ACCESS_TYPE_ID', 'ID');
   }
 }