+++ src/Models/User.php
@@ -6,8 +6,8 @@ use ItDelmax\AuthCache\Models\Branch;
 use ItDelmax\AuthCache\Models\Partner;
 use ItDelmax\AuthCache\Models\Uposljeni;
 use ItDelmax\AuthCache\Models\AccountType;
-use ItDelmax\AuthCache\Models\EtgApi;
-use ItDelmax\AuthCache\Models\EtgApiUser;
+use ItDelmax\AuthCache\Models\DmxApi;
+use ItDelmax\AuthCache\Models\DmxApiUser;
 use ItDelmax\AuthCache\Models\Traits\CachesRelationships;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -109,7 +109,7 @@ class User extends Authenticatable implements MustVerifyEmail
 
   public function apiAccess(): HasMany
   {
-    return $this->hasMany(EtgApiUser::class, 'USER_ID', 'user_id')
+    return $this->hasMany(DmxApiUser::class, 'USER_ID', 'user_id')
       ->where('IS_ACTIVE', 1)
       ->where(function ($query) {
         $query->whereNull('EXPIRES_AT')
@@ -154,7 +154,7 @@ class User extends Authenticatable implements MustVerifyEmail
 
   public function getAccessibleApis()
   {
-    return EtgApi::whereHas('apiUsers', function ($query) {
+    return DmxApi::whereHas('apiUsers', function ($query) {
       $query->where('USER_ID', $this->user_id)
         ->where('IS_ACTIVE', 1)
         ->where(function ($q) {